Implemented proper usage of liquidity Adjustment and Inventory adjustmet
git-svn-id: http://macdonald.dcs.gla.ac.uk/~tws/repositories/jagora/branches/ivo-summer-2015@81 ef8dee83-5f3d-43c0-baf0-9562671df143

+++ impl/src/main/java/uk/ac/glasgow/jagora/trader/impl/MarketMakerBasic/MarketMakerBasicBuilder.java
@@ -21,7 +21,7 @@ public class MarketMakerBasicBuilder {
     private StockWarehouse stockWarehouse;
 
     private Float marketShare;
-    private Long spread;
+    private Double spread;
 
     public MarketMakerBasicBuilder (String name) {
         this.name = name;
@@ -58,7 +58,7 @@ public class MarketMakerBasicBuilder {
         return this;
     }
 
-    public MarketMakerBasicBuilder setSpread(Long spread) {
+    public MarketMakerBasicBuilder setSpread(Double spread) {
         this.spread = spread;
         return this;
     }
